requirements.txt
streamlit>=1.28.0
requests>=2.31.0
polyline>=2.0.0
folium>=0.15.0
streamlit-folium>=0.15.0
python-dotenv>=1.0.0
openai>=1.40.0
pandas>=2.0.0
openpyxl>=3.1.0
filelock>=3.13.0
pillow>=10.0.0
streamlit-audiorec>=0.1.3
